WIP: Refactoring LinearToolButton into a more generic skinnable button
