+++ TLM/TLM/U/ButtonMouseState.cs
@@ -1,0 +1,7 @@
+namespace TrafficManager.U {
+    public enum ButtonMouseState {
+        Base,
+        Hovered,
+        MouseDown,
+    }
+}
